Avoid name clashes between Game and PartialBoard
Rename PartialBoard versions of functions like `flipTileAt` and
`isComplete` to `flipTileAtWith` and `isCompleteWith` to avoid name
clashes between modules. The name change on PartialBoard is intended to
reflect the nature of the functions as relative to a Board, as opposed
to the similar functions on a Game which are not relative.