Add functions to generate lists of Tiles consistent with Clues
allConsistentFlippedTiles returns all possible lists of Tiles that could
result from flipping the unflipped Tiles in the given list of Maybe
Tiles, that are also consistent with the given Clue and flipped Tiles.

allConsistentFlippedTileSets returns a list of sets of Tiles that are
consistent with the given Clue and flipped Tiles. This provides less
information than allConsistentFlippedTiles, however is often sufficient
to determine that a particular Position of a Board is safe or unsafe to
flip. This condensed representation also avoids the combinatorial
explosion that would result from considering all 4^25 possible Board
configurations that are consistent with a given PartialBoard and set of
Clues.