Add PartialGame and Inferences
A PartialGame represents a player's view of a Game, consisting of a
PartialBoard and some Clues. Inferences are consistent but possibly
incomplete conclusions drawn from reasoning about the PartialBoard and
Clues. The Inferences list for each Position of a Board which Tiles are
thought to be consistent with what is known about the PartialGame. This
list should be narrowed down as the Game progresses - as Tiles are
flipped, this information is added to the PartialBoard of the
PartialGame, and further reasoning about the PartialGame results in a
narrower set of Inferences.

This adds a naive and incomplete function for calculating Inferences
from a PartialGame.